Add upgrade step to update the behavior profile.
If you have installed the example content profile, you will see a
warning in the Add-ons control panel that Plone does not know how to
update this profile.  We recommend that you deactivate it and then
activate it again.

Registering an upgrade step for the example profile did not work,
because Plone still links it to the 'content' profile that is no
longer defined, so it cannot find upgrade steps.

Issue #7